Split webapp example into minimal and SPA variants
- Add a simple "Webapp example" that just exports `requestHandler` as
  the `get` handler — no lib-router, no Vite, no SPA fallback. Files
  live under the default `/assets/` root with `index.html` as directory
  index. The minimum possible wiring.
- Rename the existing complex controller from `webapp.ts` to
  `webapp-spa.ts` so the simple example can occupy `webapp.ts`. Fix the
  previously malformed `cacheControl:` block (the function declaration
  had gone missing in a prior edit).
- Add `src/main/resources/assets/{index.html,style.css}` to back the
  minimal example via `include::` directives.

The lib-router callout in the Common section already covers "lib-router
optional but handy"; the simple webapp example reaffirms that by
shipping zero router code.
