+++ src/main/resources/assets/index.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>lib-static — minimal webapp</title>
+    <link rel="stylesheet" href="style.css">
+</head>
+<body>
+    <h1>Hello from lib-static</h1>
+    <p>This page is served straight from <code>/assets/</code> by <code>requestHandler</code>.</p>
+</body>
+</html>
