+++ src/main/resources/assets/style.css
@@ -1,0 +1,8 @@
+body {
+    font-family: sans-serif;
+    margin: 2rem;
+}
+
+h1 {
+    color: tomato;
+}
